# lightpilot-backend/docs/openapi-v1.0.0-rc3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LightPilot Backend",
    "description": "Structured multi-stage intent parsing and scene semantics. Mock/unavailable results require Android HOLD.",
    "version": "1.0.0-rc3"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/analyze-scene": {
      "post": {
        "summary": "Analyze Scene",
        "operationId": "analyze_scene_api_v1_analyze_scene_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeSceneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneSemantic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parse-intent": {
      "post": {
        "summary": "Parse Intent",
        "operationId": "parse_intent_api_v1_parse_intent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseIntentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeSceneRequest": {
        "properties": {
          "frame_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Frame Id"
          },
          "intent_revision": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Intent Revision"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent"
          },
          "image_base64": {
            "type": "string",
            "maxLength": 5592440,
            "minLength": 1,
            "title": "Image Base64"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "frame_id",
          "intent_revision",
          "intent",
          "image_base64"
        ],
        "title": "AnalyzeSceneRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Intent": {
        "properties": {
          "exposure_priority": {
            "type": "string",
            "enum": [
              "subject_detail",
              "highlight_detail",
              "balanced"
            ],
            "title": "Exposure Priority"
          },
          "stability_preference": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ],
            "title": "Stability Preference"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "exposure_priority",
          "stability_preference"
        ],
        "title": "Intent"
      },
      "Metrics": {
        "properties": {
          "subject_brightness": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Brightness"
          },
          "background_brightness": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Brightness"
          },
          "highlight_clipping_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlight Clipping Ratio"
          },
          "dark_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dark Ratio"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Metrics"
      },
      "ParseIntentRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
            "title": "Request Id"
          },
          "source_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Source Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "source_text"
        ],
        "title": "ParseIntentRequest"
      },
      "ParseIntentResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "mock",
              "unavailable"
            ],
            "title": "Status"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParsedIntent"
              },
              {
                "type": "null"
              }
            ]
          },
          "ambiguities": {
            "items": {
              "type": "string",
              "maxLength": 500,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 12,
            "title": "Ambiguities"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "status",
          "reason"
        ],
        "title": "ParseIntentResponse"
      },
      "ParsedIntent": {
        "properties": {
          "weights": {
            "$ref": "#/components/schemas/StageWeights"
          },
          "exposure_priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "subject_detail",
                  "highlight_detail",
                  "balanced"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure Priority"
          },
          "motion_priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "motion_clarity",
                  "low_noise",
                  "brightness_priority",
                  "motion_balanced"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Motion Priority"
          },
          "color_priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "color_accuracy",
                  "natural_skin",
                  "atmosphere_preservation",
                  "colored_light_preservation",
                  "color_stability"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Priority"
          },
          "stability_preference": {
            "type": "string",
            "enum": [
              "normal",
              "high"
            ],
            "title": "Stability Preference"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "weights",
          "exposure_priority",
          "motion_priority",
          "color_priority",
          "stability_preference"
        ],
        "title": "ParsedIntent"
      },
      "SceneSemantic": {
        "properties": {
          "frame_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Frame Id"
          },
          "intent_revision": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Intent Revision"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "mock",
              "unavailable"
            ],
            "title": "Status"
          },
          "scene": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "indoor_even_light",
                  "indoor_mixed_light",
                  "indoor_low_light",
                  "indoor_backlit",
                  "outdoor_daylight",
                  "outdoor_backlit",
                  "night_low_light",
                  "stage_colored_light",
                  "high_contrast_other",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene"
          },
          "subject_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "person",
                  "group",
                  "display",
                  "document",
                  "object",
                  "landscape",
                  "none",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Type"
          },
          "bright_region_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "none",
                  "sky",
                  "window",
                  "display",
                  "lamp",
                  "specular_reflection",
                  "mixed",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Bright Region Type"
          },
          "colored_light": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colored Light"
          },
          "uncertainty": {
            "items": {
              "type": "string",
              "maxLength": 500,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 12,
            "title": "Uncertainty"
          },
          "uncertainty_details": {
            "items": {
              "$ref": "#/components/schemas/UncertaintyDetail"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Uncertainty Details"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "frame_id",
          "intent_revision",
          "status"
        ],
        "title": "SceneSemantic"
      },
      "StageWeights": {
        "properties": {
          "exposure": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Exposure"
          },
          "motion_noise": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Motion Noise"
          },
          "color_atmosphere": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Color Atmosphere"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "exposure",
          "motion_noise",
          "color_atmosphere"
        ],
        "title": "StageWeights"
      },
      "UncertaintyDetail": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "scene_uncertain",
              "subject_type_uncertain",
              "subject_occluded",
              "bright_region_uncertain",
              "colored_light_uncertain",
              "image_blur",
              "image_too_dark",
              "image_quality_uncertain",
              "other",
              "model_not_configured",
              "model_not_connected",
              "timeout",
              "rate_limited",
              "authentication_failed",
              "connection_failed",
              "invalid_model_response",
              "model_unavailable",
              "backend_busy",
              "internal_error"
            ],
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "enum": [
              "warning",
              "blocking"
            ],
            "title": "Severity"
          },
          "affects": {
            "items": {
              "type": "string",
              "enum": [
                "scene",
                "subject_type",
                "subject_roi",
                "bright_region_type",
                "colored_light",
                "image_quality",
                "all"
              ]
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 1,
            "title": "Affects"
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "severity",
          "affects",
          "message"
        ],
        "title": "UncertaintyDetail"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}